Escala de cor para a paleta de cores Folha

Uma escala personalizada do ggplot2 para a estética de `color` usando a paleta de cores Folha. A paleta é composta por um conjunto de nove cores distintas, projetadas para serem usadas em variáveis discretas de cor em gráficos do ggplot2.

Usage

scale_color_folha(..., na.value = "grey50")

Arguments

...

Argumentos adicionais passados para discrete_scale.

na.value

A cor a ser usada para valores ausentes. O valor padrão é "grey50".

Value

Uma função de escala do ggplot2 que aplica a paleta de cores Folha.